refactor: rewrite all src file to ts
