+++ rslib.config.ts
@@ -1,19 +1,20 @@
-import { defineConfig } from "@rslib/core";
+import { defineConfig } from '@rslib/core';
 
 export default defineConfig({
   lib: [
     {
-      syntax: ["node 20"],
-      format: "cjs",
+      syntax: ['node 20'],
+      format: 'cjs',
     },
     {
-      syntax: ["node 20"],
-      format: "esm",
+      syntax: ['node 20'],
+      format: 'esm',
+      dts: true,
     },
   ],
   source: {
     entry: {
-      index: "./src/index.js",
+      index: './src/index.ts',
     },
   },
 });
